# docs/swagger.yaml
basePath: /
definitions:
...
  models.DateRange:
    properties:
      end:
        type: string
      start:
        type: string
    type: object
...
  models.DeleteEventRequest:
    properties:
      cancellation_reason:
        type: string
      event_id:
        type: string
    required:
    - event_id
    type: object
  models.DeleteEventResponse:
    properties:
      data:
        $ref: '#/definitions/models.EventDeleteData'
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.Event:
    properties:
      additional_key:
        type: string
      created_at:
        description: Audit fields
        type: string
      custom_fields:
        additionalProperties: true
        type: object
      end_datetime:
        type: string
      event_description:
        type: string
      event_title:
        description: Event details (can override show defaults)
        type: string
      generated_at:
        description: Generation tracking
        type: string
      id:
        type: string
      is_customized:
        type: boolean
      last_synced_at:
        type: string
      length_minutes:
        type: integer
      show_id:
        type: string
      show_version:
        type: integer
      start_datetime:
        description: Timing (can override show defaults)
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.EventStatus'
        description: Event metadata
      updated_at:
        type: string
      user_id:
        type: string
      youtube_key:
        type: string
      zoom_meeting_id:
        type: string
      zoom_meeting_url:
        type: string
      zoom_passcode:
        type: string
    type: object
...
  models.EventDeleteData:
    properties:
      cancelled_at:
        type: string
      event_id:
        type: string
      status:
        $ref: '#/definitions/models.EventStatus'
    type: object
  models.EventFilters:
    properties:
      date_range:
        $ref: '#/definitions/models.DateRange'
      is_customized:
        type: boolean
      show_ids:
        items:
          type: string
        type: array
      status:
        items:
          $ref: '#/definitions/models.EventStatus'
        type: array
    type: object
  models.EventInfoData:
    properties:
      event:
        $ref: '#/definitions/models.Event'
      show_details:
        $ref: '#/definitions/models.ShowSummary'
    type: object
  models.EventListData:
    properties:
      events:
        items:
          $ref: '#/definitions/models.EventListItem'
        type: array
      pagination:
        $ref: '#/definitions/models.PaginationResponse'
    type: object
  models.EventListItem:
    properties:
      end_datetime:
        type: string
      event_title:
        type: string
      has_zoom_meeting:
        type: boolean
      id:
        type: string
      is_customized:
        type: boolean
      show_id:
        type: string
      show_name:
        type: string
      start_datetime:
        type: string
      status:
        $ref: '#/definitions/models.EventStatus'
    type: object
  models.EventListRequest:
    properties:
      filters:
        $ref: '#/definitions/models.EventFilters'
      pagination:
        $ref: '#/definitions/models.PaginationOptions'
      sort:
        $ref: '#/definitions/models.EventSortOptions'
    type: object
  models.EventListResponse:
    properties:
      data:
        $ref: '#/definitions/models.EventListData'
      error:
        type: string
      success:
        type: boolean
    type: object
  models.EventSortOptions:
    properties:
      field:
        type: string
      order:
        type: string
    type: object
  models.EventStatus:
    enum:
    - scheduled
    - live
    - completed
    - cancelled
    - postponed
    type: string
    x-enum-varnames:
    - EventStatusScheduled
    - EventStatusLive
    - EventStatusCompleted
    - EventStatusCancelled
    - EventStatusPostponed
...
  models.GetEventInfoResponse:
    properties:
      data:
        $ref: '#/definitions/models.EventInfoData'
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.MonthDay:
    properties:
      date:
        type: string
      day_number:
        type: integer
      events:
        items:
          $ref: '#/definitions/models.MonthDayEvent'
        type: array
      events_count:
        type: integer
      is_current_month:
        type: boolean
    type: object
  models.MonthDayEvent:
    properties:
      duration_minutes:
        type: integer
      id:
        type: string
      is_customized:
        type: boolean
      show_name:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/models.EventStatus'
    type: object
  models.MonthListData:
    properties:
      events_by_status:
        additionalProperties:
          type: integer
        type: object
      month:
        type: integer
      month_name:
        type: string
      timezone:
        type: string
      total_events:
        type: integer
      weeks:
        items:
          $ref: '#/definitions/models.MonthWeek'
        type: array
      year:
        type: integer
    type: object
  models.MonthListRequest:
    properties:
      filters:
        $ref: '#/definitions/models.EventFilters'
      month:
        maximum: 12
        minimum: 1
        type: integer
      timezone:
        type: string
      year:
        maximum: 2030
        minimum: 2020
        type: integer
    required:
    - month
    - year
    type: object
  models.MonthListResponse:
    properties:
      data:
        $ref: '#/definitions/models.MonthListData'
      error:
        type: string
      success:
        type: boolean
    type: object
  models.MonthWeek:
    properties:
      days:
        items:
          $ref: '#/definitions/models.MonthDay'
        type: array
      week_number:
        type: integer
    type: object
  models.PaginationOptions:
    properties:
      limit:
        type: integer
      page:
        type: integer
      sort:
        type: string
    type: object
  models.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.RepeatPattern:
    enum:
    - none
    - daily
    - weekly
    - biweekly
    - monthly
    - custom
    type: string
    x-enum-varnames:
    - RepeatNone
    - RepeatDaily
    - RepeatWeekly
    - RepeatBiweekly
    - RepeatMonthly
    - RepeatCustom
...
  models.ShowStatus:
    enum:
    - active
    - paused
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - ShowStatusActive
    - ShowStatusPaused
    - ShowStatusCompleted
    - ShowStatusCancelled
  models.ShowSummary:
    properties:
      id:
        type: string
      repeat_pattern:
        $ref: '#/definitions/models.RepeatPattern'
      show_name:
        type: string
      status:
        $ref: '#/definitions/models.ShowStatus'
    type: object
...
  models.UpdateEventRequest:
    properties:
      additional_key:
        type: string
      custom_fields:
        additionalProperties: true
        type: object
      event_description:
        type: string
      event_id:
        type: string
      event_title:
        type: string
      length_minutes:
        type: integer
      start_datetime:
        type: string
      youtube_key:
        type: string
      zoom_meeting_id:
        type: string
      zoom_meeting_url:
        type: string
      zoom_passcode:
        type: string
    required:
    - event_id
    type: object
  models.UpdateEventResponse:
    properties:
      data:
        $ref: '#/definitions/models.Event'
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.WeekDay:
    properties:
      date:
        type: string
      day_name:
        type: string
      events:
        items:
          $ref: '#/definitions/models.WeekDayEvent'
        type: array
    type: object
  models.WeekDayEvent:
    properties:
      end_time:
        type: string
      event_title:
        type: string
      id:
        type: string
      is_customized:
        type: boolean
      show_name:
        type: string
      start_time:
        type: string
      status:
        $ref: '#/definitions/models.EventStatus'
    type: object
  models.WeekListData:
    properties:
      days:
        items:
          $ref: '#/definitions/models.WeekDay'
        type: array
      timezone:
        type: string
      total_events:
        type: integer
      week_end:
        type: string
      week_start:
        type: string
    type: object
  models.WeekListRequest:
    properties:
      filters:
        $ref: '#/definitions/models.EventFilters'
      timezone:
        type: string
      week_start:
        type: string
    required:
    - week_start
    type: object
  models.WeekListResponse:
    properties:
      data:
        $ref: '#/definitions/models.WeekListData'
      error:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A Go-based microservice for planning and scheduling YouTube live streams.
    Helps content creators organize their streaming schedule and manage content.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: St. Planer - YouTube Stream Planner API
  version: "1.0"
paths:
...
  /api/v1/event/delete:
    delete:
      consumes:
      - application/json
      description: Cancel a specific event while preserving show template
      parameters:
      - description: Event deletion data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DeleteEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteEventResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel/delete event
      tags:
      - events
  /api/v1/event/info/{event_id}:
    get:
      description: Get detailed information about a specific event including show
        details
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetEventInfoResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get event details
      tags:
      - events
  /api/v1/event/list:
    post:
      consumes:
      - application/json
      description: Get paginated list of events with filtering and sorting options
      parameters:
      - description: Event list filters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.EventListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EventListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List events with filtering
      tags:
      - events
  /api/v1/event/monthList:
    post:
      consumes:
      - application/json
      description: Get events organized by weeks and days for a specific month
      parameters:
      - description: Month view request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MonthListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MonthListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get month view of events
      tags:
      - events
  /api/v1/event/update:
    put:
      consumes:
      - application/json
      description: Update specific event with custom fields while preserving show
        inheritance
      parameters:
      - description: Event update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateEventResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update event details
      tags:
      - events
  /api/v1/event/weekList:
    post:
      consumes:
      - application/json
      description: Get events organized by days for a specific week
      parameters:
      - description: Week view request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.WeekListRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.WeekListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get week view of events
      tags:
      - events
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"